Datos/requirements.txt
Flask==3.0.2
typing_extensions==4.8.0
colorama==0.4.6
beautifulsoup4==4.12.2
geopandas==0.13.2
matplotlib==3.8.0
rasterio==1.3.8
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.2.0
scipy==1.10.1
scikit-learn==1.3.2
openai==0.27.8
nltk==3.8.1
keras==3.3.3
tensorflow==2.16.1